Refactor common normalization tests into DbtIntegrationTest
