+++ .github/workflows/test-command.yml
@@ -108,6 +108,11 @@ jobs:
           name: test-reports
           path: |
             **/${{ github.event.inputs.connector }}/build/reports/tests/**/**
+            **/normalization_test_output/**/dbt_output.log
+            **/normalization_test_output/**/destination_output.log
+            **/normalization_test_output/**/build/compiled/airbyte_utils/**
+            **/normalization_test_output/**/build/run/airbyte_utils/**
+            **/normalization_test_output/**/models/generated/**
       - name: Report Status
         if: github.ref == 'refs/heads/master' && always()
         run: ./tools/status/report.sh ${{ github.event.inputs.connector }} ${{github.repository}} ${{github.run_id}} ${{steps.test.outcome}}
