+++ airbyte-integrations/bases/base-normalization/integration_tests/test_ephemeral.py
@@ -30,30 +30,21 @@ import tempfile
 from typing import Any, Dict
 
 import pytest
-from integration_tests.integration_test_utils import (
-    change_current_test_dir,
-    copy_replace,
-    dbt_run,
-    generate_profile_yaml_file,
-    random_string,
-    run_destination_process,
-    setup_postgres_db,
-    target_schema,
-    tear_down_postgres_db,
-)
+from integration_tests.dbt_integration_test import DbtIntegrationTest
 from normalization.destination_type import DestinationType
 from normalization.transform_catalog.catalog_processor import CatalogProcessor
 
 temporary_folders = set()
+dbt_test_utils = DbtIntegrationTest()
 
 
 @pytest.fixture(scope="module", autouse=True)
 def before_all_tests(request):
-    change_current_test_dir(request)
-    setup_postgres_db()
+    dbt_test_utils.change_current_test_dir(request)
+    dbt_test_utils.setup_postgres_db()
     os.environ["PATH"] = os.path.abspath("../.venv/bin/") + ":" + os.environ["PATH"]
     yield
-    tear_down_postgres_db()
+    dbt_test_utils.tear_down_postgres_db()
     for folder in temporary_folders:
         print(f"Deleting temporary test folder {folder}")
         shutil.rmtree(folder, ignore_errors=True)
@@ -61,7 +52,7 @@ def before_all_tests(request):
 
 @pytest.fixture
 def setup_test_path(request):
-    change_current_test_dir(request)
+    dbt_test_utils.change_current_test_dir(request)
     print(f"Running from: {pathlib.Path().absolute()}")
     print(f"Current PATH is: {os.environ['PATH']}")
     yield
@@ -106,27 +97,29 @@ def run_test(integration_type: str, column_count: int):
     print("Testing ephemeral")
     destination_type = DestinationType.from_string(integration_type)
     # Create the test folder with dbt project and appropriate destination settings to run integration tests from
-    test_root_dir = setup_test_dir(integration_type, "test_ephemeral")
-    destination_config = generate_profile_yaml_file(destination_type, test_root_dir)
+    test_root_dir = setup_test_dir(integration_type)
+    destination_config = dbt_test_utils.generate_profile_yaml_file(destination_type, test_root_dir)
     # generate a catalog and associated dbt models files
     generate_dbt_models(destination_type, test_root_dir, column_count)
     # Use destination connector to create empty _airbyte_raw_* tables to use as input for the test
     assert setup_input_raw_data(integration_type, test_root_dir, destination_config)
-    dbt_run(test_root_dir)
+    dbt_test_utils.dbt_run(test_root_dir)
 
 
-def setup_test_dir(integration_type: str, test_resource_name: str) -> str:
+def setup_test_dir(integration_type: str) -> str:
     """
     We prepare a clean folder to run the tests from.
     """
-    test_root_dir = tempfile.mkdtemp(dir="/tmp/", prefix="normalization_test_", suffix=f"_{integration_type.lower()}")
+    test_root_dir = (
+        f"{pathlib.Path().joinpath('..', 'build', 'normalization_test_output', integration_type.lower()).resolve()}"
+    )
+    os.makedirs(test_root_dir, exist_ok=True)
+    test_root_dir = tempfile.mkdtemp(dir=test_root_dir)
     temporary_folders.add(test_root_dir)
     shutil.rmtree(test_root_dir, ignore_errors=True)
-    os.makedirs(test_root_dir)
-    test_root_dir = f"{test_root_dir}/{test_resource_name}"
     print(f"Setting up test folder {test_root_dir}")
     shutil.copytree("../dbt-project-template", test_root_dir)
-    copy_replace("../dbt-project-template/dbt_project.yml", os.path.join(test_root_dir, "dbt_project.yml"))
+    dbt_test_utils.copy_replace("../dbt-project-template/dbt_project.yml", os.path.join(test_root_dir, "dbt_project.yml"))
     return test_root_dir
 
 
@@ -155,7 +148,7 @@ def setup_input_raw_data(integration_type: str, test_root_dir: str, destination_
         "/data/catalog.json",
     ]
     # Force a reset in destination raw tables
-    return run_destination_process("", test_root_dir, commands)
+    return dbt_test_utils.run_destination_process("", test_root_dir, commands)
 
 
 def generate_dbt_models(destination_type: DestinationType, test_root_dir: str, column_count: int):
@@ -187,9 +180,9 @@ def generate_dbt_models(destination_type: DestinationType, test_root_dir: str, c
     if column_count == 1:
         catalog_config["streams"][0]["stream"]["json_schema"]["properties"]["_airbyte_id"] = {"type": "integer"}
     else:
-        for column in [random_string(5) for _ in range(column_count)]:
+        for column in [dbt_test_utils.random_string(5) for _ in range(column_count)]:
             catalog_config["streams"][0]["stream"]["json_schema"]["properties"][column] = {"type": "string"}
     catalog = os.path.join(test_root_dir, "catalog.json")
     with open(catalog, "w") as fh:
         fh.write(json.dumps(catalog_config))
-    catalog_processor.process(catalog, "_airbyte_data", target_schema)
+    catalog_processor.process(catalog, "_airbyte_data", dbt_test_utils.target_schema)
