+++ docs/tutorials/transformation-and-normalization/transformations-with-sql.md
@@ -26,7 +26,7 @@ This could be useful if:
 
 1. you have different usage than analytics that could be handled with these initial data in raw JSON format. 
 2. you can implement your own Transformer \(even in a different language such as Java or in Spark for example, or another transformation tool: DBT or Dataform\) 
-3. you want to customize and change how the data is normalized with your own queries \(add deduplication logic since Airbyte is not doing it natively yet?\)
+3. you want to customize and change how the data is normalized with your own queries.
 
 In order to do so, we will now describe how you can leverage the basic normalization outputs that Airbyte generates to build your own transformations if you don't want to start from scratch.
